+++ internal/repository/index/json/json.go
@@ -1,4 +1,4 @@
-package indexer
+package json
 
 import (
 	"encoding/json"
